backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymupdf
chromadb
openai
python-multipart
